Finish converting math module;

+++ src/api/types/curve.c
@@ -59,7 +59,7 @@ static int l_lovrCurveGetPointCount(lua_State* L) {
 static int l_lovrCurveGetPoint(lua_State* L) {
   Curve* curve = luax_checktype(L, 1, Curve);
   int index = luaL_checkinteger(L, 2) - 1;
-  lovrAssert(index >= 0 && index < lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
+  lovrAssert(index >= 0 && index < (int) lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
   float point[3];
   lovrCurveGetPoint(curve, index, point);
   lua_pushnumber(L, point[0]);
@@ -71,7 +71,7 @@ static int l_lovrCurveGetPoint(lua_State* L) {
 static int l_lovrCurveSetPoint(lua_State* L) {
   Curve* curve = luax_checktype(L, 1, Curve);
   int index = luaL_checkinteger(L, 2) - 1;
-  lovrAssert(index >= 0 && index < lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
+  lovrAssert(index >= 0 && index < (int) lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
   float point[3];
   luax_readvec3(L, 3, point, NULL);
   lovrCurveSetPoint(curve, index, point);
@@ -82,7 +82,7 @@ static int l_lovrCurveAddPoint(lua_State* L) {
   Curve* curve = luax_checktype(L, 1, Curve);
   float point[3];
   int i = luax_readvec3(L, 2, point, NULL);
-  int index = lua_isnoneornil(L, i) ? lovrCurveGetPointCount(curve) : luaL_checkinteger(L, i) - 1;
+  usize index = lua_isnoneornil(L, i) ? lovrCurveGetPointCount(curve) : luaL_checkinteger(L, i) - 1;
   lovrAssert(index >= 0 && index <= lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
   lovrCurveAddPoint(curve, point, index);
   return 0;
@@ -91,7 +91,7 @@ static int l_lovrCurveAddPoint(lua_State* L) {
 static int l_lovrCurveRemovePoint(lua_State* L) {
   Curve* curve = luax_checktype(L, 1, Curve);
   int index = luaL_checkinteger(L, 2) - 1;
-  lovrAssert(index >= 0 && index < lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
+  lovrAssert(index >= 0 && index < (int) lovrCurveGetPointCount(curve), "Invalid Curve point index: %d", index + 1);
   lovrCurveRemovePoint(curve, index);
   return 0;
 }
